% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{bhuvad_theme}
\alias{bhuvad_theme}
\title{Custom theme}
\usage{
bhuvad_theme(rl = 1.1)
}
\arguments{
\item{rl}{a numeric, scaling factor to apply to text sizes}
}
\value{
a ggplot2 theme
}
\description{
Custom theme
}
\examples{

p1 = ggplot2::ggplot()
p1 + bhuvad_theme()

}
