% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFrequencies.R
\name{computeMsigWordFreq}
\alias{computeMsigWordFreq}
\title{Compute word frequencies for a single MSigDB collection}
\usage{
computeMsigWordFreq(
  msigGsc,
  weight = NULL,
  measure = c("tfidf", "tf"),
  rmwords = getMsigBlacklist()
)
}
\arguments{
\item{msigGsc}{a GeneSetCollection object, containing gene sets from the
MSigDB. The \code{\link[GSEABase:getObjects]{GSEABase::getBroadSets()}} function can be used to parse XML
files downloaded from MSigDB.}

\item{weight}{a named numeric vector, containing weights to apply to each
gene-set. This can be -log10(FDR), -log10(p-value) or an enrichment score
(ideally unsigned).}

\item{measure}{a character, specifying how frequencies should be computed.
"tf" uses term frequencies and "tfidf" (default) applies inverse document
frequency weights to term frequencies.}

\item{rmwords}{a character vector, containing a blacklist of words to discard
from the analysis.}
}
\value{
a list, containing two data.frames summarising the results of the
frequency analysis on gene set names and short descriptions.
}
\description{
Compute word frequencies for a single MSigDB collection
}
\examples{
data(hgsc)
freq <- computeMsigWordFreq(hgsc, measure = 'tfidf')

}
