% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{current_theme}
\alias{current_theme}
\title{Custom theme}
\usage{
current_theme(rl = 1.1)
}
\arguments{
\item{rl}{a numeric, scaling factor to apply to text sizes}
}
\value{
a ggplot2 theme
}
\description{
Custom theme
}
\examples{

}
