% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFrequencies.R
\name{getMsigBlacklist}
\alias{getMsigBlacklist}
\title{Blacklist words for MSigDB gene set text mining}
\usage{
getMsigBlacklist(custom = c())
}
\arguments{
\item{custom}{a character vector, containing list of words to add onto
existing blacklist.}
}
\value{
a character vector, containing list of blacklist works
}
\description{
List of words to discard when performing text mining MSigDB gene set names
and short descriptions.
}
\examples{

getMsigBlacklist('blacklist')

}
