% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFrequencies.R
\name{getMsigExclusionList}
\alias{getMsigExclusionList}
\title{Exclusion list of words for MSigDB gene set text mining}
\usage{
getMsigExclusionList(custom = c())
}
\arguments{
\item{custom}{a character vector, containing list of words to add onto
existing exclusion list.}
}
\value{
a character vector, containing words to be excluded from the text mining analysis.
}
\description{
List of words to discard when performing text mining MSigDB gene set names
and short descriptions.
}
\examples{
getMsigExclusionList('remove')

}
