% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetNetworks.R
\name{getMsigNeighbour}
\alias{getMsigNeighbour}
\title{Get similar gene sets from overlap networks}
\usage{
getMsigNeighbour(srcsig, ig, thresh = 0.1)
}
\arguments{
\item{srcsig}{a character, naming the gene set to begin the search with.}

\item{ig}{an igraph object, containing a network of gene set overlaps
computed using \code{\link[=computeMsigNetwork]{computeMsigNetwork()}}.}

\item{thresh}{a numeric, specifying the threshold to discard pairs of gene
sets (default is 0.1).}
}
\value{
a character, containing the names of gene sets that overlap with the
source signature.
}
\description{
Get similar gene sets from overlap networks
}
\examples{

data("msigOverlapNetwork")
neighbours <- getMsigNeighbour('HALLMARK_HYPOXIA', msigOverlapNetwork, 0.1)

}
