% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichnets.R
\docType{data}
\name{hgsc}
\alias{hgsc}
\title{The Hallmark collection from the MSigDB}
\format{
A GeneSetCollection object with 50 GeneSet objects representing the
50 Hallmark gene expression signatures.
}
\usage{
hgsc
}
\description{
The molecular signatures database (MSigDB) is a collection of over 25000 gene
expression signatures. Signatures in v7.1 are divided into 8 categories. The
Hallmarks collection contains gene expression signatures representing
molecular processes that are hallmarks in cancer development and progression.
}
\references{
Subramanian, A., Tamayo, P., Mootha, V. K., Mukherjee, S., Ebert,
B. L., Gillette, M. A., ... & Mesirov, J. P. (2005). Gene set enrichment
analysis: a knowledge-based approach for interpreting genome-wide
expression profiles. Proceedings of the National Academy of Sciences,
102(43), 15545-15550.

Liberzon, A., Subramanian, A., Pinchback, R., Thorvaldsdóttir, H., Tamayo,
P., & Mesirov, J. P. (2011). Molecular signatures database (MSigDB) 3.0.
Bioinformatics, 27(12), 1739-1740.

Liberzon, A., Birger, C., Thorvaldsdóttir, H., Ghandi, M., Mesirov, J. P.,
& Tamayo, P. (2015). The molecular signatures database hallmark gene set
collection. Cell systems, 1(6), 417-425.
}
\keyword{datasets}
