% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vissE.R
\docType{data}
\name{mem_mat_hs}
\alias{mem_mat_hs}
\title{Binary membership matrix for the Human MSigDB}
\format{
A dgCMatrix (sparse) object, with gene sets along the rows and Entrez
IDs along the columns.
}
\usage{
mem_mat_hs
}
\description{
This object stores the Human molecular signatures database (MSigDB) in binary
format as a membership matrix. Gene signatures are along the rows and Entrez
IDs are along the columns.
}
\keyword{datasets}
