% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vissE.R
\docType{data}
\name{mem_mat_mm}
\alias{mem_mat_mm}
\title{Binary membership matrix for the Mouse MSigDB}
\format{
A dgCMatrix (sparse) object, with gene sets along the rows and Entrez
IDs along the columns.
}
\usage{
mem_mat_mm
}
\description{
This object stores the Mouse molecular signatures database (MSigDB) in binary
format as a membership matrix. Gene signatures are along the rows and Entrez
IDs are along the columns.
}
\keyword{datasets}
