% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichnets.R
\docType{data}
\name{msigOverlapNetwork}
\alias{msigOverlapNetwork}
\title{An overlap network of signatures in the MSigDB}
\format{
An igraph object, containing the overlap network computed using all
non-empty gene expression signatures in MSigDB.
}
\usage{
msigOverlapNetwork
}
\description{
The molecular signatures database (MSigDB) is a collection of over 25000 gene
expression signatures. Signatures in v7.1 are divided into 8 categories. This
network represents the precomputed overlap between all non-empty signatures
network represents the pre-computed overlap between all non-empty signatures
in the \code{msigdb} data. The network was computed using \code{\link[=computeMsigOverlap]{computeMsigOverlap()}}
and \code{\link[=computeMsigNetwork]{computeMsigNetwork()}} functions with default parameters.
}
\keyword{datasets}
