% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFunctions.R
\name{plotMsigNetwork}
\alias{plotMsigNetwork}
\title{Plot a gene set overlap network}
\usage{
plotMsigNetwork(
  ig,
  markGroups = NULL,
  enrichStat = NULL,
  nodeSF = 1,
  edgeSF = 1,
  lytFunc = igraph::layout_with_graphopt,
  lytParams = list()
)
}
\arguments{
\item{ig}{an igraph object, containing a network of gene set overlaps
computed using \code{\link[=computeMsigNetwork]{computeMsigNetwork()}}.}

\item{markGroups}{a named list, of character vectors or numeric indices
specifying node groupings. Each element of the list represent a group and
contains a character vector with node names. Up to 12 groups can be
visualised in the plot.}

\item{nodeSF}{a numeric, indicating the scaling factor to apply to node
sizes.}

\item{edgeSF}{a numeric, indicating the scaling factor to apply to edge
widths.}

\item{lytFunc}{a function, that computes layouts and returns a matrix with 2
columns specifying the x and y coordinates of nodes. Layout functions in
the igraph package can be used here.}

\item{lytParams}{a named list, containing additional parameters to be passed
on to the layout function.}
}
\value{
a ggplot2 object
}
\description{
Plots a network of gene set overlap with overlap computed using the
\code{\link[=computeMsigOverlap]{computeMsigOverlap()}} and a graph created using \code{\link[=computeMsigNetwork]{computeMsigNetwork()}}.
}
\examples{

data(hgsc)
ovlap <- computeMsigOverlap(hgsc)
ig <- computeMsigNetwork(ovlap, hgsc)
groups <- list('g1' = c(1, 10, 11, 14, 21), 'g2' = c(3, 5, 13))

plotMsigNetwork(ig, markGroups = groups)

}
