% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characteriseGeneset.R
\name{characteriseGeneset}
\alias{characteriseGeneset}
\title{Functionally characterise a list of genes}
\usage{
characteriseGeneset(gs, thresh = 0.15, measure = c("ovlapcoef", "jaccard"))
}
\arguments{
\item{gs}{a GeneSet object, representing the list of genes that need to be
characterised.}

\item{thresh}{a numeric, specifying the threshold to discard pairs of gene
sets (default 0.15).}

\item{measure}{a character, specifying the similarity measure to use:
\code{jaccard} for the Jaccard Index and \code{ovlapcoef} for the Overlap
Coefficient.}
}
\value{
an igraph object, containing gene sets that are similar to the query
set. The network contains relationships between results of the query too.
}
\description{
This function can be used to perform a network-based enrichment analysis of a
list of genes. The list of genes are characterised based on their similarity
with gene sets from the MSigDB. A network of similar gene sets is retrieved
using this function.
}
\examples{
data(hgsc)

#create a geneset using one of the Hallmark gene sets
mySet <- GSEABase::GeneSet(GSEABase::geneIds(hgsc[[2]]), setName = 'MySet')

\dontrun{
#characterise the custom gene set
ig <- characteriseGeneset(mySet)
plotMsigNetwork(ig)
}

}
