% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetNetworks.R
\name{computeMsigOverlap}
\alias{computeMsigOverlap}
\title{Compute gene set overlap}
\usage{
computeMsigOverlap(msigGsc, thresh = 0.1, measure = c("jaccard", "ovlapcoef"))
}
\arguments{
\item{msigGsc}{a GeneSetCollection object.}

\item{thresh}{a numeric, specifying the threshold to discard pairs of gene
sets.}

\item{measure}{a character, specifying the similarity measure to use:
\code{jaccard} for the Jaccard Index and \code{ovlapcoef} for the Overlap
Coefficient.}
}
\value{
a data.frame, containing the overlap structure of gene sets
represented as a network in the simple interaction format (SIF)
}
\description{
Compute overlap between gene sets from a GeneSetCollection using the Jaccard
index or the overlap coefficient. These values can then be used to compute a
network of gene set overlaps.
}
\examples{

data(hgsc)
ovlap <- computeMsigOverlap(hgsc)

}
