% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordFrequencies.R
\name{computeMsigWordFreq}
\alias{computeMsigWordFreq}
\title{Compute word frequencies for a single MSigDB collection}
\usage{
computeMsigWordFreq(
  msigGsc,
  measure = c("tf", "tfidf"),
  rmwords = getMsigBlacklist()
)
}
\arguments{
\item{msigGsc}{a GeneSetCollection object, containing gene sets from the
MSigDB. The \code{\link[GSEABase:getObjects]{GSEABase::getBroadSets()}} function can be used to parse XML
files downloaded from MSigDB.}

\item{measure}{a character, specifying how frequencies should be computed.
"tf" (default) uses term frequencies and "tfidf" applies inverse document
frequency weights to term frequencies.}

\item{rmwords}{a character vector, containing a blacklist of words to discard
from the analysis.}
}
\value{

}
\description{
Compute word frequencies for a single MSigDB collection
}
\examples{

data(hgsc)
freq <- computeMsigWordFreq(hgsc, measure = 'tf')

}
